backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==77.0.3
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
